# perf: Batch nested Parquet decoding