--- a/crates/polars-parquet/src/arrow/read/deserialize/binview/nested.rs
+++ b/crates/polars-parquet/src/arrow/read/deserialize/binview/nested.rs
@@ -1,27 +1,50 @@
 use std::collections::VecDeque;
 
-use arrow::array::{ArrayRef, MutableBinaryViewArray};
+use arrow::array::{ArrayRef, MutableBinaryViewArray, View};
 use arrow::bitmap::MutableBitmap;
 use arrow::datatypes::ArrowDataType;
 use polars_error::PolarsResult;
-use polars_utils::iter::FallibleIterator;
 
-use crate::parquet::page::{DataPage, DictPage};
-use crate::read::deserialize::binary::decoders::{
-    build_nested_state, deserialize_plain, BinaryDict, BinaryNestedState,
-};
+use crate::parquet::encoding::hybrid_rle::DictionaryTranslator;
+use crate::parquet::encoding::Encoding;
+use crate::parquet::error::ParquetResult;
+use crate::parquet::page::{split_buffer, DataPage, DictPage};
+use crate::read::deserialize::binary::decoders::{deserialize_plain, BinaryDict, ValuesDictionary};
+use crate::read::deserialize::binary::utils::BinaryIter;
 use crate::read::deserialize::binview::basic::finish;
 use crate::read::deserialize::nested_utils::{next, NestedDecoder};
-use crate::read::deserialize::utils::MaybeNext;
+use crate::read::deserialize::utils::{
+    binary_views_dict, not_implemented, page_is_filtered, page_is_optional, MaybeNext, PageState,
+};
 use crate::read::{InitNested, NestedState, PagesIter};
 
-#[derive(Debug, Default)]
-struct BinViewDecoder {}
+#[derive(Debug)]
+pub(crate) struct State<'a> {
+    pub is_optional: bool,
+    pub translation: StateTranslation<'a>,
+}
+
+#[derive(Debug)]
+pub(crate) enum StateTranslation<'a> {
+    Unit(BinaryIter<'a>),
+    Dictionary(ValuesDictionary<'a>, Option<Vec<View>>),
+}
+
+impl<'a> PageState<'a> for State<'a> {
+    fn len(&self) -> usize {
+        match &self.translation {
+            StateTranslation::Unit(iter) => iter.size_hint().0,
+            StateTranslation::Dictionary(values, _) => values.len(),
+        }
+    }
+}
+
+struct BinViewDecoder;
 
 type DecodedStateTuple = (MutableBinaryViewArray<[u8]>, MutableBitmap);
 
 impl<'a> NestedDecoder<'a> for BinViewDecoder {
-    type State = BinaryNestedState<'a>;
+    type State = State<'a>;
     type Dictionary = BinaryDict;
     type DecodedState = DecodedStateTuple;
 
@@ -30,7 +53,29 @@ impl<'a> NestedDecoder<'a> for BinViewDecoder {
         page: &'a DataPage,
         dict: Option<&'a Self::Dictionary>,
     ) -> PolarsResult<Self::State> {
-        build_nested_state(page, dict)
+        let is_optional = page_is_optional(page);
+        let is_filtered = page_is_filtered(page);
+
+        if is_filtered {
+            return Err(not_implemented(page));
+        }
+
+        let translation = match (page.encoding(), dict) {
+            (Encoding::PlainDictionary | Encoding::RleDictionary, Some(dict)) => {
+                StateTranslation::Dictionary(ValuesDictionary::try_new(page, dict)?, None)
+            },
+            (Encoding::Plain, _) => {
+                let values = split_buffer(page)?.values;
+                let values = BinaryIter::new(values, page.num_values());
+                StateTranslation::Unit(values)
+            },
+            _ => return Err(not_implemented(page)),
+        };
+
+        Ok(State {
+            is_optional,
+            translation,
+        })
     }
 
     fn with_capacity(&self, capacity: usize) -> Self::DecodedState {
@@ -40,51 +85,43 @@ impl<'a> NestedDecoder<'a> for BinViewDecoder {
         )
     }
 
-    fn push_valid(
+    fn push_n_valid(
         &self,
         state: &mut Self::State,
         decoded: &mut Self::DecodedState,
-    ) -> PolarsResult<()> {
+        n: usize,
+    ) -> ParquetResult<()> {
         let (values, validity) = decoded;
-        match state {
-            BinaryNestedState::Optional(page) => {
-                let value = page.next().unwrap_or_default();
-                values.push_value_ignore_validity(value);
-                validity.push(true);
+        match &mut state.translation {
+            StateTranslation::Unit(page) => {
+                // @TODO: This should probably be optimized to a better loop
+                for value in page.by_ref().take(n) {
+                    values.push_value_ignore_validity(value);
+                }
             },
-            BinaryNestedState::Required(page) => {
-                let value = page.next().unwrap_or_default();
-                values.push_value_ignore_validity(value);
-            },
-            BinaryNestedState::RequiredDictionary(page) => {
-                let dict_values = &page.dict;
-                let item = page
-                    .values
-                    .next()
-                    .map(|index| dict_values.value(index as usize))
-                    .unwrap_or_default();
-                values.push_value_ignore_validity(item);
-                page.values.get_result()?;
-            },
-            BinaryNestedState::OptionalDictionary(page) => {
-                let dict_values = &page.dict;
-                let item = page
-                    .values
-                    .next()
-                    .map(|index| dict_values.value(index as usize))
-                    .unwrap_or_default();
-                values.push_value_ignore_validity(item);
-                validity.push(true);
-                page.values.get_result()?;
+            StateTranslation::Dictionary(page, views_dict) => {
+                let views_dict =
+                    views_dict.get_or_insert_with(|| binary_views_dict(values, page.dict));
+                let translator = DictionaryTranslator(views_dict);
+                page.values
+                    .translate_and_collect_n_into(values.views_mut(), n, &translator)?;
+                if let Some(validity) = values.validity() {
+                    validity.extend_constant(n, true);
+                }
             },
         }
+
+        if state.is_optional {
+            validity.extend_constant(n, true);
+        }
+
         Ok(())
     }
 
-    fn push_null(&self, decoded: &mut Self::DecodedState) {
+    fn push_n_nulls(&self, decoded: &mut Self::DecodedState, n: usize) {
         let (values, validity) = decoded;
-        values.push_null();
-        validity.push(false);
+        values.extend_null(n);
+        validity.extend_constant(n, false);
     }
 
     fn deserialize_dict(&self, page: &DictPage) -> Self::Dictionary {
@@ -134,7 +171,7 @@ impl<I: PagesIter> Iterator for NestedIter<I> {
                 &mut self.remaining,
                 &self.init,
                 self.chunk_size,
-                &BinViewDecoder::default(),
+                &BinViewDecoder,
             );
             match maybe_state {
                 MaybeNext::Some(Ok((nested, decoded))) => {